Fixing Functional test project
